requirements.txt
django>=4.0,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary>=2.9,<3.0

